:recycle: Extract snackbar logic to useSnackbar hook
